Neue Syscalls rein geholt

+++ Betriebssystem Source-Code/apps/animation/src/main.rs
@@ -29,7 +29,7 @@ pub fn main() {
     // Daten holen
     let pid = usr_get_pid();
     let mut namebuffer: [u8; BUFFERLENGH] = [0; BUFFERLENGH];
-    usr_read_process_name(namebuffer.as_mut_ptr(), BUFFERLENGH as u64) as usize;
+    usr_read_process_name(namebuffer.as_mut_ptr(), BUFFERLENGH) as usize;
     let actual_name: &str = unsafe {
         from_utf8_unchecked(
             namebuffer
@@ -42,6 +42,7 @@ pub fn main() {
 
     // Ausgabe
     print_setpos!(50, 36, "Name: {}; pid: {}", actual_name, pid);
+
 
     // Animation
     animate_charmander(500, 400);
